FM: Select band based on country
Checks the locale country code, and if found in the lists of
known codes, returns the appropriate regional band. If not,
accepts the default from resources as before.

Change-Id: Icaa98ded11e548afc8be446e7bc80a1066bee873
Issue-id: CYNGNOS-738

FM: cleaned up band selection code

Based on the code reviews on my previous commit, my earlier
code (which was admittedly slapped together in a hurry) was
too sloppy. this is a more elegant solution.

Change-Id: I08fc64ad628a3c8cb9e634310b89ac8abae2d7c8
Issue-id: CYNGNOS-738
Signed-off-by: Adithya R <radithya2002@gmail.com>